+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.0] - (07.02.2023)
+## Changes
+- added possibility expose Contact Sensor in HomeKit app for Ports and SSIDs to use with automations
+- config.schema updated
+- bump dependencies
+- cleanup
+
 ## [0.13.10] - (31.12.2022)
 ## Changes
 - bump dependencies